Move widget display from DefaultController to WidgetController
