+++ src/AppBundle/Controller/WidgetController.php
@@ -2,6 +2,7 @@
 
 namespace AppBundle\Controller;
 
+use AppBundle\Entity\ShiftBucket;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
@@ -17,6 +18,44 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
  */
 class WidgetController extends Controller
 {
+
+    /**
+     * Widget display
+     * 
+     * @Route("/", name="widget", methods={"GET"})
+     */
+    public function widgetAction(Request $request)
+    {
+        $job_id = $request->get('job_id');
+        $buckets = array();
+        $display_end = $request->query->has('display_end') ? ($request->get('display_end') == 1) : false;
+        $display_on_empty = $request->query->has('display_on_empty') ? ($request->get('display_on_empty') == 1) : false;
+        $title = $request->query->has('title') ? ($request->get('title') == 1) : true;
+        $job = null;
+        if ($job_id) {
+            $em = $this->getDoctrine()->getManager();
+            $job = $em->getRepository('AppBundle:Job')->find($job_id);
+            if ($job) {
+                $shifts = $em->getRepository('AppBundle:Shift')->findFuturesWithJob($job);
+                foreach ($shifts as $shift) {
+                    $day = $shift->getStart()->format("d m Y");
+                    $interval = $shift->getIntervalCode();
+                    if (!isset($buckets[$interval . $day])) {
+                        $buckets[$interval . $day] = new ShiftBucket();
+                    }
+                    $buckets[$interval . $day]->addShift($shift);
+                }
+            }
+        }
+
+        return $this->render('widget/widget.html.twig', [
+            'job' => $job,
+            'buckets' => $buckets,
+            'display_end' => $display_end,
+            'display_on_empty' => $display_on_empty,
+            'title' => $title
+        ]);
+    }
 
     /**
      * Widget generator
